uint32 > uint64 fix

+++ contracts/MultiResource_EIP/interfaces/IMultiResource.sol
@@ -8,39 +8,39 @@ pragma solidity ^0.8.0;
 interface IMultiResource is IERC721 {
 
     struct Resource {
-        uint32 id; //8 bytes
+        uint64 id; //8 bytes
         string metadataURI; //32+
-        uint64[] custom;
+        uint128[] custom;
     }
 
-    event ResourceSet(uint32 resourceId);
+    event ResourceSet(uint64 resourceId);
 
-    event ResourceAddedToToken(uint256 indexed tokenId, uint32 resourceId);
+    event ResourceAddedToToken(uint256 indexed tokenId, uint64 resourceId);
 
-    event ResourceAccepted(uint256 indexed tokenId, uint32 resourceId);
+    event ResourceAccepted(uint256 indexed tokenId, uint64 resourceId);
 
-    event ResourceRejected(uint256 indexed tokenId, uint32 resourceId);
+    event ResourceRejected(uint256 indexed tokenId, uint64 resourceId);
 
     event ResourcePrioritySet(uint256 indexed tokenId);
 
     event ResourceOverwriteProposed(
         uint256 indexed tokenId,
-        uint32 resourceId,
-        uint32 overwrites
+        uint64 resourceId,
+        uint64 overwrites
     );
 
-    event ResourceOverwritten(uint256 indexed tokenId, uint32 overwritten);
+    event ResourceOverwritten(uint256 indexed tokenId, uint64 overwritten);
 
-    event ResourceCustomDataSet(uint32 resourceId, uint64 customResourceId);
+    event ResourceCustomDataSet(uint64 resourceId, uint128 customResourceId);
 
     event ResourceCustomDataAdded(
-        uint32 resourceId,
-        uint64 customResourceId
+        uint64 resourceId,
+        uint128 customResourceId
     );
 
     event ResourceCustomDataRemoved(
-        uint32 resourceId,
-        uint64 customResourceId
+        uint64 resourceId,
+        uint128 customResourceId
     );
 
     function acceptResource(uint256 tokenId, uint256 index) external;
@@ -53,11 +53,11 @@ interface IMultiResource is IERC721 {
 
     function getActiveResources(
         uint256 tokenId
-    ) external view returns(uint32[] memory);
+    ) external view returns(uint64[] memory);
 
     function getPendingResources(
         uint256 tokenId
-    ) external view returns(uint32[] memory);
+    ) external view returns(uint64[] memory);
 
     function getActiveResourcePriorities(
         uint256 tokenId
@@ -65,16 +65,16 @@ interface IMultiResource is IERC721 {
 
     function getResourceOverwrites(
         uint256 tokenId,
-        uint32 resourceId
-    ) external view returns(uint32);
+        uint64 resourceId
+    ) external view returns(uint64);
 
     function getResource(
-        uint32 resourceId
+        uint64 resourceId
     ) external view returns (Resource memory);
 
     function getCustomResourceData(
-        uint32 resourceId,
-        uint64 customResourceId
+        uint64 resourceId,
+        uint128 customResourceId
     ) external view returns (bytes memory);
 
     function tokenURI(
